+++ backend/__test__/virusTotal.test.js
@@ -27,5 +27,4 @@ describe('virusTotal API', function() {
         done();
       });
   });
-});
-
+});;
