+++ backend/middleware/authMiddleware.js
