--- db.config.js
@@ -1,9 +1,0 @@
-// db.config.js stores database connection details
-module.exports = {
-  HOST: 'localhost',
-  USER: 'root',
-  PASSWORD: 'olecko',
-  DB: 'safepassdb',
-  dialect: 'mysql',
-};
-
